refactor: reorganise shared components/utils to fit existing layout
