+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2021-01-25T14:58:44.611Z\n"
-"PO-Revision-Date: 2021-01-25T14:58:44.611Z\n"
+"POT-Creation-Date: 2021-01-26T12:03:08.275Z\n"
+"PO-Revision-Date: 2021-01-26T12:03:08.275Z\n"
 
 msgid "Are you sure you want to cancel? Unsaved changes will be lost"
 msgstr ""
@@ -18,6 +18,9 @@ msgid "Yes, cancel"
 msgstr ""
 
 msgid "Data set"
+msgstr ""
+
+msgid "Delete selected"
 msgstr ""
 
 msgid ""
@@ -134,6 +137,27 @@ msgid "Clickatell"
 msgstr ""
 
 msgid "Generic"
+msgstr ""
+
+msgid "Show"
+msgstr ""
+
+msgid "per page"
+msgstr ""
+
+msgid "Viewing {{firstItem}}-{{lastItem}} of {{total}}"
+msgstr ""
+
+msgid "Previous"
+msgstr ""
+
+msgid "Page"
+msgstr ""
+
+msgid "of {{pageCount}}"
+msgstr ""
+
+msgid "Next"
 msgstr ""
 
 msgid "Program"
@@ -277,9 +301,6 @@ msgstr ""
 msgid "User group"
 msgstr ""
 
-msgid "Delete selected"
-msgstr ""
-
 msgid "Something went wrong whilst loading gateways"
 msgstr ""
 
@@ -361,27 +382,6 @@ msgid "Open logs of all SMS messages received by DHIS2."
 msgstr ""
 
 msgid "View all received SMS"
-msgstr ""
-
-msgid "Show"
-msgstr ""
-
-msgid "per page"
-msgstr ""
-
-msgid "Viewing {{firstItem}}-{{lastItem}} of {{total}}"
-msgstr ""
-
-msgid "Previous"
-msgstr ""
-
-msgid "Page"
-msgstr ""
-
-msgid "of {{pageCount}}"
-msgstr ""
-
-msgid "Next"
 msgstr ""
 
 msgid "Filter by status"
