+++ src/utils/createSearchString.js
@@ -1,4 +1,4 @@
-export default query => {
+export const createSearchString = query => {
     const queryString = Object.keys(query)
         .reduce((acc, paramKey) => {
             const paramValue = query[paramKey]
