+++ src/utils/index.js
@@ -1,1 +1,2 @@
 export * from './get'
+export * from './createSearchString.js'
