+++ src/views/sent_sms_list/SentSmsList.js
@@ -1,9 +1,9 @@
 import { NoticeBox, CenteredContent, CircularLoader } from '@dhis2/ui'
 import { useDataQuery } from '@dhis2/app-runtime'
 import React, { useState, useEffect } from 'react'
-import useQueryParams from '../useQueryParams'
+import { useQueryParams } from '../../hooks'
 import { PageHeadline } from '../../headline'
-import DeleteSelectedButton from '../delete_selected_button/DeleteSelectedButton'
+import DeleteSelectedButton from '../../delete_selected_button/DeleteSelectedButton'
 import i18n from '../../locales'
 import SmsTable from './SmsTable'
 import StatusFilter from './StatusFilter'
